+++ logdissectlib/__init__.py
@@ -1,0 +1,5 @@
+__version__ = '1.4-dev'
+__author__ = 'Dan Persons <dpersonsdev@gmail.com>'
+__license__ = 'MIT License'
+__github__ = 'https://github.com/dogoncouch/logdissect'
+__all__ = ['core']