+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change log for [logdissect](https://github.com/dogoncouch/logdissect)
 ## [Unreleased]
 ### Added
 - 'inpcap' parser for parsing pcap files
+###Moved
+- parsers moved to 'logdissectlib' to allow external use
 
 ## [1.3.1] - 2017-04-24
 ### Fixed